Rewrite gridUncertainty
gridUncertainty now creates different scenarios systematically by sliding a set distance as dictated by the number of splits on a single grid. The output has also been updated to reflect the change, and now reports the amount moved in x and y driection which lead to the grid scenario(s) with the smallest AOO.
A raster layer of the grids with the smallest AOO is now included within the output.