Add gridUncertainty functions
Two additional gridUncertainty functions have been added, and the base
function has been renamed to gridUncertatinyBase. gridUncertainty now
calculates the AOOs for the input distribution continuously until the
AOO does not improve after a certain rounds/splits.
gridUncertaintySimulation can be used to perform this manually, with a
set amount of simulations to take place. Currently uses lots of for
loops, can this be improved?